requirements.txt
Jinja2==3.1.4
MarkupSafe==2.1.5
Nuitka==2.3.11
numpy==2.0.0
ordered-set==4.1.0
pyqtgraph==0.13.7
pyserial==3.5
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
qt-material==2.14
shiboken6==6.7.2
zstandard==0.22.0
